pyproject.toml
[project]
name = "dspy-tutorials"
version = "0.1.0"
description = "Me coding along various DSPy tutorials"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.0.3",
    "faiss-cpu>=1.12.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "ujson>=5.11.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "types-ujson>=5.10.0.20250822",
]
